refactor(useTeamValues): Move to own file
