+++ packages/client/src/modules/play/context/hooks/shared/index.ts
@@ -1,1 +1,2 @@
+export { useTeamValues } from "./useTeamValues";
 export { useTeamValuesForTeam } from "./useTeamValuesForTeam";
