+++ packages/client/src/modules/play/context/hooks/shared/useTeamValuesForTeam.ts
@@ -1,10 +1,7 @@
-import { range } from "lodash";
 import { useMemo } from "react";
 import { TeamValues, usePersonaByUserId, usePlay } from "../../playContext";
 import { mean } from "../../../../../lib/math";
-import { GameStepType, isStepOfType } from "../../../constants";
-import { IGame, Player } from "../../../../../utils/types";
-import { sumAllValues } from "../../../../persona";
+import { buildStepToData } from "./utils";
 
 export { useTeamValuesForTeam };
 
@@ -85,35 +82,4 @@ function useTeamValuesForTeam({ teamId }: { teamId?: number }): {
   return {
     teamValues,
   };
-}
-
-function buildStepToData(
-  dataType: GameStepType,
-  game: IGame,
-  players: Player[],
-  personaByUserId: ReturnType<typeof usePersonaByUserId>
-) {
-  return Object.fromEntries(
-    range(0, game.lastFinishedStep + 1)
-      .filter((step) => isStepOfType(step, dataType))
-      .map((step: number) => [
-        step,
-        buildStepData(dataType, step, players, personaByUserId),
-      ])
-  );
-}
-
-function buildStepData(
-  dataType: GameStepType,
-  step: number,
-  players: Player[],
-  personaByUserId: ReturnType<typeof usePersonaByUserId>
-) {
-  return mean(
-    players
-      .map((p) => personaByUserId[p.userId].getPersonaAtStep(step)[dataType])
-      .map((data) =>
-        parseInt(sumAllValues(data as { type: string; value: number }[]))
-      )
-  );
 }
